# v3_api-docs.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Domicare API",
      "description": "API documentation for Domicare application",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Generated server url"
      }
    ],
    "paths": {
      "/users": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getUsers",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 20
              }
            },
            {
              "name": "sortBy",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "sortDirection",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "default": "asc"
              }
            },
            {
              "name": "spec",
              "in": "query",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/SpecificationUser"
              }
            },
            {
              "name": "pageable",
              "in": "query",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/Pageable"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ResultPagingDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "user-controller"
          ],
          "operationId": "updateUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/users/roles": {
        "put": {
          "tags": [
            "user-controller"
          ],
          "operationId": "updateRoleForUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRoleForUserRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/users/{id}": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getUserById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "user-controller"
          ],
          "operationId": "deleteUserById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "UpdateUserRequest": {
          "required": [
            "id"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "email": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            }
          }
        },
        "RoleDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "active": {
              "type": "boolean"
            },
            "createBy": {
              "type": "string"
            },
            "updateBy": {
              "type": "string"
            },
            "createAt": {
              "type": "string",
              "format": "date-time"
            },
            "updateAt": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "UserDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "avatar": {
              "type": "string"
            },
            "googleId": {
              "type": "string"
            },
            "emailConfirmationToken": {
              "type": "string"
            },
            "createBy": {
              "type": "string"
            },
            "updateBy": {
              "type": "string"
            },
            "createAt": {
              "type": "string",
              "format": "date-time"
            },
            "updateAt": {
              "type": "string",
              "format": "date-time"
            },
            "roles": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RoleDTO"
              }
            },
            "emailConfirmed": {
              "type": "boolean"
            }
          }
        },
        "UpdateRoleForUserRequest": {
          "required": [
            "roleIds",
            "userId"
          ],
          "type": "object",
          "properties": {
            "userId": {
              "type": "integer",
              "format": "int64"
            },
            "roleIds": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        },
...
        "SpecificationUser": {
          "type": "object"
        },
        "Pageable": {
          "type": "object",
          "properties": {
            "page": {
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            },
            "size": {
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            },
            "sort": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "Meta": {
          "type": "object",
          "properties": {
            "page": {
              "type": "integer",
              "format": "int32"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "total": {
              "type": "integer",
              "format": "int64"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "ResultPagingDTO": {
          "type": "object",
          "properties": {
            "meta": {
              "$ref": "#/components/schemas/Meta"
            },
            "data": {
              "type": "object"
            }
          }
        },
...
      }
    }
  }